Separate styles from rest of the code
