+++ src/components/App.js
@@ -1,4 +1,4 @@
-import './App.css';
+import '../styles/App.css';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import Home from '../container/Home';
