+++ src/components/Card.js
@@ -1,0 +1,7 @@
+const Card = () => (
+  <div className="card">
+    <h1>This is a card</h1>
+  </div>
+);
+
+export default Card;
