+++ src/container/Home.js
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
+import Card from '../components/Card';
 
 import { fetchCrypto } from '../redux/crypto/cryptoActions';
 
@@ -9,13 +10,16 @@ const Home = () => {
     dispact(fetchCrypto());
   }, []);
   const state = useSelector(state => state);
-  const { crypto } = state.cryptoReducer;
+  const { crypto, loading } = state.cryptoReducer;
+  console.log(loading);
   return (
     <>
+      <Card />
       {crypto.map(coin => {
         const { asset_id: id, name, price_usd: price } = coin;
         return (
           <div key={id}>
+            <p>{id}</p>
             <h2>{name}</h2>
             <h3>{price}</h3>
           </div>
