+++ src/styles/index.css
@@ -29,3 +29,6 @@ code {
     'Courier New',
     monospace;
 }
+.container{
+  background-color: red;
+}
